@@ -13,12 +13,16 @@
         {"name": "Johannes Schmitt", "email": "schmittjoh@gmail.com"}
     ],
     "require": {
-        "php": "~5.5|~7.0"
+        "php": "~7.1"
     },
     "require-dev": {
-        "phpunit/phpunit":         "~4.8|~5.0",
-        "satooshi/php-coveralls":  "~0.6",
-        "predis/predis":           "~1.0"
+        "alcaeus/mongo-php-adapter": "^1.1",
+        "mongodb/mongodb": "^1.1",
+        "phpunit/phpunit": "^5.7",
+        "predis/predis":   "~1.0"
+    },
+    "suggest": {
+        "alcaeus/mongo-php-adapter": "Required to use legacy MongoDB driver"
     },
     "conflict": {
         "doctrine/common": ">2.2,<2.4"
@@ -31,7 +35,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.6.x-dev"
+            "dev-master": "1.7.x-dev"
         }
     }
 }
