@@ -37,8 +37,9 @@ class GoogleAnalyticsVisibilityPages extends ProcessPluginBase implements Contai
   protected $migrationPlugin;
 
   /**
-   * Whether or not to skip Google Analytics that use PHP for visibility. Only
-   * applies if the PHP module is not enabled.
+   * Whether or not to skip Google Analytics that use PHP for visibility.
+   *
+   * Only applies if the PHP module is not enabled.
    *
    * @var bool
    */
@@ -61,12 +62,12 @@ class GoogleAnalyticsVisibilityPages extends ProcessPluginBase implements Contai
    * {@inheritdoc}
    */
   public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition, MigrationInterface $migration = NULL) {
-    $migration_configuration = array(
-      'migration' => array(
+    $migration_configuration = [
+      'migration' => [
         'd6_user_role',
         'd7_user_role',
-      ),
-    );
+      ],
+    ];
     return new static(
       $configuration,
       $plugin_id,
