@@ -48,12 +48,12 @@ class GoogleAnalyticsVisibilityRoles extends ProcessPluginBase implements Contai
    * {@inheritdoc}
    */
   public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition, MigrationInterface $migration = NULL) {
-    $migration_configuration = array(
-      'migration' => array(
+    $migration_configuration = [
+      'migration' => [
         'd6_user_role',
         'd7_user_role',
-      ),
-    );
+      ],
+    ];
     return new static(
       $configuration,
       $plugin_id,
@@ -72,7 +72,7 @@ class GoogleAnalyticsVisibilityRoles extends ProcessPluginBase implements Contai
     // Remove role IDs disabled in D6/D7.
     $roles = array_filter($roles);
 
-    $user_role_roles = array();
+    $user_role_roles = [];
 
     if ($roles) {
       foreach ($roles as $key => $role_id) {
