@@ -129,15 +129,15 @@ class MetatagDisplayExtender extends DisplayExtenderPluginBase {
    * This output is returned as an array.
    */
   public function optionsSummary(&$categories, &$options) {
-    $categories['metatags'] = array(
+    $categories['metatags'] = [
       'title' => t('Meta tags'),
       'column' => 'second',
-    );
-    $options['metatags'] = array(
+    ];
+    $options['metatags'] = [
       'category' => 'metatags',
       'title' => t('Meta tags'),
       'value' => $this->hasMetatags() ? t('Overridden') : t('Using defaults'),
-    );
+    ];
   }
 
   /**
@@ -165,7 +165,7 @@ class MetatagDisplayExtender extends DisplayExtenderPluginBase {
    *   The meta tag values.
    */
   public function getMetatags() {
-    $metatags = array();
+    $metatags = [];
 
     if (!empty($this->options['metatags'])) {
       $metatags = $this->options['metatags'];
