@@ -5,7 +5,7 @@
 
 (function ($, window) {
 
-  "use strict";
+  'use strict';
 
   /**
    * Provide the summary information for the tracking settings vertical tabs.
