@@ -5,9 +5,9 @@
 
 (function ($, Drupal, drupalSettings) {
 
-  /*eslint no-console:0*/
+  /* eslint no-console: 0, max-nested-callbacks: ["error", 4] */
 
-  "use strict";
+  'use strict';
 
   Drupal.google_analytics = {};
 
@@ -15,12 +15,12 @@
 
     // Attach mousedown, keyup, touchstart events to document only and catch
     // clicks on all elements.
-    $(document.body).on("mousedown keyup touchstart", function (event) {
-      console.group("Running Google Analytics for Drupal.");
+    $(document.body).on('mousedown keyup touchstart', function (event) {
+      console.group('Running Google Analytics for Drupal.');
       console.info("Event '%s' has been detected.", event.type);
 
       // Catch the closest surrounding link of a clicked element.
-      $(event.target).closest("a,area").each(function () {
+      $(event.target).closest('a,area').each(function () {
         console.info("Closest element '%o' has been found. URL '%s' extracted.", this, this.href);
 
         // Is the clicked URL internal?
@@ -28,28 +28,28 @@
           // Skip 'click' tracking, if custom tracking events are bound.
           if ($(this).is('.colorbox') && (drupalSettings.google_analytics.trackColorbox)) {
             // Do nothing here. The custom event will handle all tracking.
-            console.info("Click on .colorbox item has been detected.");
+            console.info('Click on .colorbox item has been detected.');
           }
           // Is download tracking activated and the file extension configured
           // for download tracking?
           else if (drupalSettings.google_analytics.trackDownload && Drupal.google_analytics.isDownload(this.href)) {
             // Download link clicked.
             console.info("Download url '%s' has been found. Tracked download as extension '%s'.", Drupal.google_analytics.getPageUrl(this.href), Drupal.google_analytics.getDownloadExtension(this.href).toUpperCase());
-            ga("send", {
-              "hitType": "event",
-              "eventCategory": "Downloads",
-              "eventAction": Drupal.google_analytics.getDownloadExtension(this.href).toUpperCase(),
-              "eventLabel": Drupal.google_analytics.getPageUrl(this.href),
-              "transport": "beacon"
+            ga('send', {
+              hitType: 'event',
+              eventCategory: 'Downloads',
+              eventAction: Drupal.google_analytics.getDownloadExtension(this.href).toUpperCase(),
+              eventLabel: Drupal.google_analytics.getPageUrl(this.href),
+              transport: 'beacon'
             });
           }
           else if (Drupal.google_analytics.isInternalSpecial(this.href)) {
             // Keep the internal URL for Google Analytics website overlay intact.
             console.info("Click on internal special link '%s' has been tracked.", Drupal.google_analytics.getPageUrl(this.href));
-            ga("send", {
-              "hitType": "pageview",
-              "page": Drupal.google_analytics.getPageUrl(this.href),
-              "transport": "beacon"
+            ga('send', {
+              hitType: 'pageview',
+              page: Drupal.google_analytics.getPageUrl(this.href),
+              transport: 'beacon'
             });
           }
           else {
@@ -61,24 +61,24 @@
           if (drupalSettings.google_analytics.trackMailto && $(this).is("a[href^='mailto:'],area[href^='mailto:']")) {
             // Mailto link clicked.
             console.info("Click on e-mail '%s' has been tracked.", this.href.substring(7));
-            ga("send", {
-              "hitType": "event",
-              "eventCategory": "Mails",
-              "eventAction": "Click",
-              "eventLabel": this.href.substring(7),
-              "transport": "beacon"
+            ga('send', {
+              hitType: 'event',
+              eventCategory: 'Mails',
+              eventAction: 'Click',
+              eventLabel: this.href.substring(7),
+              transport: 'beacon'
             });
           }
           else if (drupalSettings.google_analytics.trackOutbound && this.href.match(/^\w+:\/\//i)) {
             if (drupalSettings.google_analytics.trackDomainMode !== 2 || (drupalSettings.google_analytics.trackDomainMode === 2 && !Drupal.google_analytics.isCrossDomain(this.hostname, drupalSettings.google_analytics.trackCrossDomains))) {
               // External link clicked / No top-level cross domain clicked.
               console.info("Outbound link '%s' has been tracked.", this.href);
-              ga("send", {
-                "hitType": "event",
-                "eventCategory": "Outbound links",
-                "eventAction": "Click",
-                "eventLabel": this.href,
-                "transport": "beacon"
+              ga('send', {
+                hitType: 'event',
+                eventCategory: 'Outbound links',
+                eventAction: 'Click',
+                eventLabel: this.href,
+                transport: 'beacon'
               });
             }
             else {
@@ -95,9 +95,9 @@
     if (drupalSettings.google_analytics.trackUrlFragments) {
       window.onhashchange = function () {
         console.info("Track URL '%s' as pageview. Hash '%s' has changed.", location.pathname + location.search + location.hash, location.hash);
-        ga("send", {
-          "hitType": "pageview",
-          "page": location.pathname + location.search + location.hash
+        ga('send', {
+          hitType: 'pageview',
+          page: location.pathname + location.search + location.hash
         });
       };
     }
@@ -105,13 +105,13 @@
     // Colorbox: This event triggers when the transition has completed and the
     // newly loaded content has been revealed.
     if (drupalSettings.google_analytics.trackColorbox) {
-      $(document).on("cbox_complete", function () {
-        var href = $.colorbox.element().attr("href");
+      $(document).on('cbox_complete', function () {
+        var href = $.colorbox.element().attr('href');
         if (href) {
           console.info("Colorbox transition to url '%s' has been tracked.", Drupal.google_analytics.getPageUrl(href));
-          ga("send", {
-            "hitType": "pageview",
-            "page": Drupal.google_analytics.getPageUrl(href)
+          ga('send', {
+            hitType: 'pageview',
+            page: Drupal.google_analytics.getPageUrl(href)
           });
         }
       });
@@ -142,7 +142,7 @@
    * @return {boolean} isDownload
    */
   Drupal.google_analytics.isDownload = function (url) {
-    var isDownload = new RegExp("\\.(" + drupalSettings.google_analytics.trackDownloadExtensions + ")([\?#].*)?$", "i");
+    var isDownload = new RegExp('\\.(' + drupalSettings.google_analytics.trackDownloadExtensions + ')([\?#].*)?$', 'i');
     return isDownload.test(url);
   };
 
@@ -155,7 +155,7 @@
    * @return {boolean} isInternal
    */
   Drupal.google_analytics.isInternal = function (url) {
-    var isInternal = new RegExp("^(https?):\/\/" + window.location.host, "i");
+    var isInternal = new RegExp('^(https?):\/\/' + window.location.host, 'i');
     return isInternal.test(url);
   };
 
@@ -171,7 +171,7 @@
    * @return {boolean} isInternalSpecial
    */
   Drupal.google_analytics.isInternalSpecial = function (url) {
-    var isInternalSpecial = new RegExp("(\/go\/.*)$", "i");
+    var isInternalSpecial = new RegExp('(\/go\/.*)$', 'i');
     return isInternalSpecial.test(url);
   };
 
@@ -189,7 +189,7 @@
    *   Internal website URL.
    */
   Drupal.google_analytics.getPageUrl = function (url) {
-    var extractInternalUrl = new RegExp("^(https?):\/\/" + window.location.host, "i");
+    var extractInternalUrl = new RegExp('^(https?):\/\/' + window.location.host, 'i');
     return url.replace(extractInternalUrl, '');
   };
 
@@ -200,10 +200,10 @@
    *   The web url to check.
    *
    * @return {string} getDownloadExtension
-   *   The file extension of the passed url. e.g. "zip", "txt"
+   *   The file extension of the passed url. e.g. 'zip', 'txt'
    */
   Drupal.google_analytics.getDownloadExtension = function (url) {
-    var extractDownloadextension = new RegExp("\\.(" + drupalSettings.google_analytics.trackDownloadExtensions + ")([\?#].*)?$", "i");
+    var extractDownloadextension = new RegExp('\\.(' + drupalSettings.google_analytics.trackDownloadExtensions + ')([\?#].*)?$', 'i');
     var extension = extractDownloadextension.exec(url);
     return (extension === null) ? '' : extension[1];
   };
