@@ -73,7 +73,7 @@ class GoogleAnalyticsSearchTest extends WebTestBase {
 
     // Save the node.
     $this->drupalPostForm('node/add/page', $edit, t('Save'));
-    $this->assertRaw(t('@type %title has been created.', ['@type' => 'Basic page', '%title' => $edit['title[0][value]']]), 'Basic page created.');
+    $this->assertText(t('@type @title has been created.', ['@type' => 'Basic page', '@title' => $edit['title[0][value]']]), 'Basic page created.');
 
     // Index the node or it cannot found.
     $this->cronRun();
@@ -83,7 +83,7 @@ class GoogleAnalyticsSearchTest extends WebTestBase {
     $this->assertRaw('window.google_analytics_search_results = 1;', '[testGoogleAnalyticsSearch]: One search result found.');
 
     $this->drupalPostForm('node/add/page', $edit, t('Save'));
-    $this->assertRaw(t('@type %title has been created.', ['@type' => 'Basic page', '%title' => $edit['title[0][value]']]), 'Basic page created.');
+    $this->assertText(t('@type @title has been created.', ['@type' => 'Basic page', '@title' => $edit['title[0][value]']]), 'Basic page created.');
 
     // Index the node or it cannot found.
     $this->cronRun();
