@@ -11,9 +11,7 @@ use Drupal\serialization\Normalizer\NormalizerBase;
 class MetatagNormalizer extends NormalizerBase {
 
   /**
-   * The interface or class that this Normalizer supports.
-   *
-   * @var string
+   * {@inheritdoc}}
    */
   protected $supportedInterfaceOrClass = 'Drupal\metatag\Plugin\Field\MetatagEntityFieldItemList';
 
@@ -23,13 +21,18 @@ class MetatagNormalizer extends NormalizerBase {
    * @see metatag_get_tags_from_route();
    */
   public function normalize($field_item, $format = NULL, array $context = []) {
-
     $entity = $field_item->getEntity();
 
     $tags = metatag_get_tags_from_route($entity);
 
     foreach ($tags['#attached']['html_head'] as $tag) {
-      $normalized['value'][$tag[1]] = $tag[0]['#attributes']['content'];
+      // @todo Work out a proper, long-term fix for this.
+      if (isset($tag[0]['#attributes']['content'])) {
+        $normalized['value'][$tag[1]] = $tag[0]['#attributes']['content'];
+      }
+      elseif (isset($tag[0]['#attributes']['href'])) {
+        $normalized['value'][$tag[1]] = $tag[0]['#attributes']['href'];
+      }
     }
 
     if (isset($context['langcode'])) {
