@@ -10,7 +10,7 @@ use \Drupal\metatag\Plugin\metatag\Tag\MetaNameBase;
  * @MetatagTag(
  *   id = "geo_position",
  *   label = @Translation("Geographical position"),
- *   description = @Translation("Geo-spatial information in 'latitude, longitude' format, e.g. '50.167958, -97.133185'; <a href='http://en.wikipedia.org/wiki/ICBM'>see Wikipedia for details</a>."),
+ *   description = @Translation("Geo-spatial information in 'latitude, longitude' format, e.g. '50.167958, -97.133185'; <a href='https://en.wikipedia.org/wiki/Geographic_coordinate_system'>see Wikipedia for details</a>."),
  *   name = "geo.position",
  *   group = "advanced",
  *   weight = 0,
