@@ -1,34 +1,16 @@
 <?php
 
+namespace Drupal\metatag\Plugin\metatag\Tag;
+
 /**
  * This base plugin allows "itemprop"-style meta tags, e.g. Google plus tags, to
  * be further customized.
  */
-
-namespace Drupal\metatag\Plugin\metatag\Tag;
-
 abstract class MetaItempropBase extends MetaNameBase {
+
   /**
-   * Display the meta tag.
+   * {@inheritDoc}
    */
-  public function output() {
-    if (empty($this->value)) {
-      // If there is no value, we don't want a tag output.
-      $element = '';
-    }
-    else {
-      // Parse out the image URL, if needed.
-      $value = $this->parseImageURL();
+  protected $name_attribute = 'itemprop';
 
-      $element = [
-        '#tag' => 'meta',
-        '#attributes' => [
-          'itemprop' => $this->name,
-          'content' => $value,
-        ]
-      ];
-    }
-
-    return $element;
-  }
 }
